T4A3. Use inheritance for reuse benefits
